Check there are IDs before trying to use an IN()

+++ system/cms/modules/streams_core/src/Pyro/Module/Streams_core/FieldAssignmentModel.php
@@ -158,6 +158,10 @@ class FieldAssignmentModel extends FieldModel
     {
         $field_ids = FieldModel::all()->modelKeys();
 
+        if (! $field_ids) {
+            return true;
+        }
+
         return static::whereNotIn('field_id', $field_ids)->delete();
     }
